Add files via upload
 Problem Statement- Al Based Allocation Engine

You have a Pool of Employees of different age groups who want to plan their investments for retirement. Each Employee has a different target for their retirement savings and can take different levels of risk towards their investments for retirements There is an option of 6 stocks (S1.,S2,S3,S4,S5,S6) in which employees con invest their money periodically for Retirement. Some stocks are less risky but may offer average profits even if market is doing well, while some stocks are highly risky but may offer great profits in good market conditions

Assume you also have access to market data of all other stocks listed on exchange (Synthesized data can  be used from Kaggle or other

You are required to create an Al based algorithm to allocate the contribution from employees in these & stocks depending on attributes of Employee, attributes of the stock or any other attributes you might want to consider in the environment.

The Allocation should be published in percentage as is outcome of the algorithm for Example for Employee E1 the Algorithm may allocate 10% in S1.20% in S2 and 70% in S3 During the Hackathon you may be required to change your algorithm on the thy, so please be ready to show your algorithm through Jupyter Notebook or any other Python editor as your feel appropriate                    
Which Kaggle dataset we should use for meeting the requirement of that algorithm.